Mana capping

+++ Project Hardlight/Assets/Scripts/FighterStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Data", menuName = "Battle/FighterStats", order = 1)]
 public class FighterStats : ScriptableObject
 {
-    public float maxHealth = 100;
-    public float maxMana = 10;
+    public int maxHealth = 100;
+    public int maxMana = 10;
     public float movementSpeed = 1.0f;
 }
